changed regtest configuration to include activation of forks
