+++ rskj-core/src/main/java/org/ethereum/config/blockchain/regtest/RegTestConfig.java
@@ -17,14 +17,14 @@
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package org.ethereum.config.blockchain;
+package org.ethereum.config.blockchain.regtest;
 
 import co.rsk.config.BridgeConstants;
 import co.rsk.config.BridgeRegTestConstants;
 import co.rsk.core.BlockDifficulty;
+import org.ethereum.config.blockchain.GenesisConfig;
 
 import java.math.BigInteger;
-
 
 public class RegTestConfig extends GenesisConfig {
 
@@ -60,7 +60,7 @@ public class RegTestConfig extends GenesisConfig {
         public byte getChainId() {
             return RegTestConstants.CHAIN_ID;
         }
-    };
+    }
 
     public RegTestConfig() {
         super(new RegTestConstants());
@@ -68,26 +68,6 @@ public class RegTestConfig extends GenesisConfig {
 
     @Override
     public boolean areBridgeTxsFree() {
-        return true;
-    }
-
-    @Override
-    public boolean isRfs50() {
-        return true;
-    }
-
-    @Override
-    public boolean isRfs55() {
-        return true;
-    }
-
-    @Override
-    public boolean isRfs94() {
-        return true;
-    }
-
-    @Override
-    public boolean isRcs230() {
         return true;
     }
 }
