+++ rskj-core/src/main/java/org/ethereum/config/net/AbstractNetConfig.java
@@ -23,6 +23,8 @@ import org.ethereum.config.BlockchainConfig;
 import org.ethereum.config.BlockchainNetConfig;
 import org.ethereum.config.Constants;
 
+import java.util.Arrays;
+
 /**
  * Created by Anton Nashatyrev on 25.02.2016.
  */
