+++ rskj-core/src/test/java/co/rsk/remasc/RemascContractExecuteTest.java
@@ -22,7 +22,7 @@ import co.rsk.config.RemascConfig;
 import co.rsk.config.RemascConfigFactory;
 import co.rsk.config.TestSystemProperties;
 import org.ethereum.config.BlockchainNetConfig;
-import org.ethereum.config.blockchain.RegTestConfig;
+import org.ethereum.config.blockchain.regtest.RegTestConfig;
 import org.ethereum.vm.PrecompiledContracts;
 import org.ethereum.vm.program.Program;
 import org.junit.BeforeClass;
